norme + master stack

+++ srcs/gui/colorutils.c
@@ -14,10 +14,10 @@
 
 static Uint32			avercolor(Uint32 *tbl)
 {
-	Uint32				r;
-	Uint32				g;
-	Uint32				b;
-	unsigned int		i;
+	unsigned long long	r;
+	unsigned long long	g;
+	unsigned long long	b;
+	unsigned long long	i;
 
 	i = 0;
 	r = 0;
